+++ src/pm/hydra/pm/pmiserv/Makefile.mk
@@ -5,15 +5,6 @@
 
 AM_CPPFLAGS += -I$(top_srcdir)/pm/utils
 
-noinst_HEADERS +=              \
-    pm/pmiserv/common.h        \
-    pm/pmiserv/pmi_v2_common.h \
-    pm/pmiserv/pmip.h          \
-    pm/pmiserv/pmip_pmi.h      \
-    pm/pmiserv/pmiserv.h       \
-    pm/pmiserv/pmiserv_pmi.h   \
-    pm/pmiserv/pmiserv_utils.h
-
 bin_PROGRAMS += hydra_pmi_proxy
 
 hydra_pmi_proxy_SOURCES = pm/pmiserv/pmip.c \
