+++ src/pm/hydra/pm/pmiserv/pmip_pmi.h
@@ -7,7 +7,7 @@
 #define PMIP_PMI_H_INCLUDED
 
 #include "hydra.h"
-#include "common.h"
+#include "pmiserv_common.h"
 
 /* PMI-1 specific definitions */
 extern struct HYD_pmcd_pmip_pmi_handle *HYD_pmcd_pmip_pmi_v1;
