+++ src/pm/hydra/pm/pmiserv/pmiserv.h
@@ -6,7 +6,7 @@
 #ifndef PMISERV_H_INCLUDED
 #define PMISERV_H_INCLUDED
 
-#include "common.h"
+#include "pmiserv_common.h"
 
 HYD_status HYD_pmcd_pmiserv_proxy_init_cb(int fd, HYD_event_t events, void *userp);
 HYD_status HYD_pmcd_pmiserv_control_listen_cb(int fd, HYD_event_t events, void *userp);
